requirements.txt
openai
pdf2image
pillow
python-dotenv
streamlit